// types/types.ts
export type form = "alcove" | "panel_van" | "integrated" | "semi_integrated";

export type transmission = "automatic" | "manual";
export type engine = "diesel" | "petrol" | "hybrid" | "electric";

export type amenities =
  | "ac"
  | "bathroom"
  | "kitchen"
  | "tv"
  | "radio"
  | "refrigerator"
  | "microwave"
  | "gas"
  | "water";